# src/server/docs/swagger.yaml
basePath: /
definitions:
...
  main.ControllerNodesResponse:
    properties:
      nodes:
        items:
          $ref: '#/definitions/model.ConnectedNode'
        type: array
    type: object
...
  main.ControllerSummaryResponse:
    properties:
      connected_nodes:
        example: 4
        type: integer
      current_offset:
        example: 12
        type: integer
      ping_requests:
        example: 8
        type: integer
      ping_results:
        example: 16
        type: integer
      status:
        example: ok
        type: string
    type: object
...
  main.ErrorResponse:
    properties:
      error:
        example: request not found
        type: string
    type: object
...
  model.ConnectedNode:
    properties:
      last_contact:
        type: string
      location:
        $ref: '#/definitions/model.Location'
      node_id:
        type: string
      organization:
        type: string
      role:
        $ref: '#/definitions/model.NodeRole'
    type: object
  model.Location:
    properties:
      lat:
        type: number
      lon:
        type: number
    type: object
...
  model.NodeRole:
    enum:
    - worker
    - controller
    - hybrid
    type: string
    x-enum-varnames:
    - NodeRoleWorker
    - NodeRoleController
    - NodeRoleHybrid
...
info:
  contact: {}
  description: HTTP API for distributed ping nodes and controller coordination.
  title: Distributed Ping API
  version: "1.0"
paths:
...
  /api/controller:
    get:
      description: Returns controller counters and current request offset.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.ControllerSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Controller summary
      tags:
      - controller
  /api/controller/nodes:
    get:
      description: Returns nodes currently known by the controller.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.ControllerNodesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: List connected controller nodes
      tags:
      - controller
...
schemes:
- http
- https
swagger: "2.0"