# src/server/routes.go
package main

// routes.go contains the HTTP API surface for both node-facing and
// controller-facing endpoints. Its only responsibility is request handling.

import (
	"net/http"
	"strconv"
	"strings"
	"time"

	"github.com/bearpro/distributed-ping/controller"
	"github.com/bearpro/distributed-ping/model"
	"github.com/gin-gonic/gin"
)

type routeHandlers struct {
	app Application
}

func registerRoutes(router *gin.Engine, app Application) {
	handlers := routeHandlers{app: app}

	router.GET("/healthz", handlers.healthz)
	router.GET("/api/application", handlers.getApplication)
	router.GET("/api/node", handlers.getNode)
	router.GET("/api/node/ping-requests", handlers.listNodePingRequests)
	router.POST("/api/node/ping-requests", handlers.createNodePingRequest)
	router.GET("/api/node/ping-requests/:id/results", handlers.getNodePingRequestResults)
	router.GET("/api/controller", handlers.getController)
	router.GET("/api/controller/nodes", handlers.listControllerNodes)
	router.GET("/api/controller/ping-requests", handlers.listControllerPingRequests)
	router.POST("/api/controller/ping-requests", handlers.createControllerPingRequest)
	router.POST("/api/controller/ping-results", handlers.createControllerPingResult)
	router.GET("/api/controller/ping-requests/:id/results", handlers.getControllerPingRequestResults)
}

// healthz godoc
//
//	@Summary		Health check
//	@Description	Returns backend health status.
//	@Tags			system
//	@Produce		json
//	@Success		200	{object}	HealthResponse
//	@Router			/healthz [get]
func (h routeHandlers) healthz(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{"status": "ok"})
}

// getApplication godoc
//
//	@Summary		Application snapshot
//	@Description	Returns the full in-memory application snapshot.
//	@Tags			application
//	@Produce		json
//	@Success		200	{object}	ApplicationSnapshot
//	@Router			/api/application [get]
func (h routeHandlers) getApplication(c *gin.Context) {
	c.JSON(http.StatusOK, h.app.Snapshot())
}

// getNode godoc
//
//	@Summary		Node summary
//	@Description	Returns local node metadata and queue counters.
//	@Tags			node
//	@Produce		json
//	@Success		200	{object}	NodeSummaryResponse
//	@Router			/api/node [get]
func (h routeHandlers) getNode(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{
		"node":                h.app.Node.Config.Identity,
		"is_controller":       h.app.Config.IsController,
		"upstream_controller": h.app.Config.ControllerURL,
		"current_offset":      h.app.Node.CurrentOffset(),
		"submitted_requests":  len(h.app.Node.SubmittedRequests()),
	})
}

// listNodePingRequests godoc
//
//	@Summary		List submitted node requests
//	@Description	Returns ping requests submitted by the current node.
//	@Tags			node
//	@Produce		json
//	@Success		200	{object}	NodePingRequestsResponse
//	@Router			/api/node/ping-requests [get]
func (h routeHandlers) listNodePingRequests(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{
		"requests": h.app.Node.SubmittedRequests(),
	})
}

// createNodePingRequest godoc
//
//	@Summary		Create node ping request
//	@Description	Creates a ping request through the configured upstream controller.
//	@Tags			node
//	@Accept			json
//	@Produce		json
//	@Param			body	body		CreateNodePingRequestBody	true	"Ping request payload"
//	@Success		201		{object}	PingRequestResponse
//	@Failure		400		{object}	ErrorResponse
//	@Failure		502		{object}	ErrorResponse
//	@Router			/api/node/ping-requests [post]
func (h routeHandlers) createNodePingRequest(c *gin.Context) {
	var body struct {
		Target string `json:"target" binding:"required"`
	}
	if err := c.ShouldBindJSON(&body); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	req, err := h.app.Node.CreatePingRequest(c.Request.Context(), body.Target, randomID("req"))
	if err != nil {
		c.JSON(http.StatusBadGateway, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusCreated, gin.H{"request": req})
}

// getNodePingRequestResults godoc
//
//	@Summary		Get node ping request results
//	@Description	Refreshes and returns results for a submitted node ping request.
//	@Tags			node
//	@Produce		json
//	@Param			id	path		string	true	"Ping request ID"
//	@Success		200	{object}	PingResultsResponse
//	@Failure		404	{object}	ErrorResponse
//	@Failure		502	{object}	ErrorResponse
//	@Router			/api/node/ping-requests/{id}/results [get]
func (h routeHandlers) getNodePingRequestResults(c *gin.Context) {
	results, err := h.app.Node.RefreshRequestResults(c.Request.Context(), c.Param("id"))
	if err != nil {
		status := http.StatusBadGateway
		if strings.Contains(err.Error(), "unknown") {
			status = http.StatusNotFound
		}
		c.JSON(status, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, gin.H{"results": results})
}

// getController godoc
//
//	@Summary		Controller summary
//	@Description	Returns controller counters and current request offset.
//	@Tags			controller
//	@Produce		json
//	@Success		200	{object}	ControllerSummaryResponse
//	@Failure		400	{object}	ErrorResponse
//	@Router			/api/controller [get]
func (h routeHandlers) getController(c *gin.Context) {
	if !ensureController(c, h.app.Controller) {
		return
	}

	nodesCount, requestsCount, resultsCount := h.app.Controller.Counts()
	c.JSON(http.StatusOK, gin.H{
		"status":          "ok",
		"current_offset":  h.app.Controller.CurrentOffset(),
		"connected_nodes": nodesCount,
		"ping_requests":   requestsCount,
		"ping_results":    resultsCount,
	})
}

// listControllerNodes godoc
//
//	@Summary		List connected controller nodes
//	@Description	Returns nodes currently known by the controller.
//	@Tags			controller
//	@Produce		json
//	@Success		200	{object}	ControllerNodesResponse
//	@Failure		400	{object}	ErrorResponse
//	@Router			/api/controller/nodes [get]
func (h routeHandlers) listControllerNodes(c *gin.Context) {
	if !ensureController(c, h.app.Controller) {
		return
	}

	c.JSON(http.StatusOK, gin.H{"nodes": h.app.Controller.NodesSnapshot()})
}

// listControllerPingRequests godoc
//
//	@Summary		List controller ping requests
//	@Description	Returns controller-managed ping requests using offset pagination.
//	@Tags			controller
//	@Produce		json
//	@Param			from_offset			query		int		false	"Read offset"	default(0)
//	@Param			limit				query		int		false	"Page size"	default(100)
//	@Param			X-Node-ID			header		string	false	"Node ID for heartbeat registration"
//	@Param			X-Node-Lat			header		number	false	"Node latitude"
//	@Param			X-Node-Lon			header		number	false	"Node longitude"
//	@Param			X-Node-Organization	header		string	false	"Node organization"
//	@Param			X-Node-Role			header		string	false	"Node role"
//	@Success		200					{object}	ControllerPingRequestsResponse
//	@Failure		400					{object}	ErrorResponse
//	@Router			/api/controller/ping-requests [get]
func (h routeHandlers) listControllerPingRequests(c *gin.Context) {
	if !ensureController(c, h.app.Controller) {
		return
	}

	if identity, ok := identityFromHeaders(c); ok {
		h.app.Controller.RegisterNode(identity, time.Now().UTC())
	}

	fromOffset, _ := strconv.Atoi(c.DefaultQuery("from_offset", "0"))
	limit, _ := strconv.Atoi(c.DefaultQuery("limit", "100"))
	requests, nextOffset := h.app.Controller.RequestsFrom(fromOffset, limit)
	c.JSON(http.StatusOK, gin.H{
		"requests":       requests,
		"next_offset":    nextOffset,
		"current_offset": h.app.Controller.CurrentOffset(),
	})
}

// createControllerPingRequest godoc
//
//	@Summary		Create controller ping request
//	@Description	Creates a ping request directly on the controller and dispatches it locally.
//	@Tags			controller
//	@Accept			json
//	@Produce		json
//	@Param			body	body		CreateControllerPingRequestBody	true	"Ping request payload"
//	@Success		201		{object}	PingRequestResponse
//	@Failure		400		{object}	ErrorResponse
//	@Router			/api/controller/ping-requests [post]
func (h routeHandlers) createControllerPingRequest(c *gin.Context) {
	if !ensureController(c, h.app.Controller) {
		return
	}

	var body struct {
		Target       string `json:"target" binding:"required"`
		SourceNodeID string `json:"source_node_id"`
	}
	if err := c.ShouldBindJSON(&body); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	req := h.app.Controller.CreateRequest(body.Target, body.SourceNodeID, time.Now().UTC(), randomID("req"))
	h.app.Node.ProcessRequest(req)
	c.JSON(http.StatusCreated, gin.H{"request": req})
}

// createControllerPingResult godoc
//
//	@Summary		Submit controller ping result
//	@Description	Stores a ping result for an existing controller request.
//	@Tags			controller
//	@Accept			json
//	@Produce		json
//	@Param			body	body		model.PingResult	true	"Ping result payload"
//	@Success		202		{object}	AcceptedResponse
//	@Failure		400		{object}	ErrorResponse
//	@Failure		404		{object}	ErrorResponse
//	@Router			/api/controller/ping-results [post]
func (h routeHandlers) createControllerPingResult(c *gin.Context) {
	if !ensureController(c, h.app.Controller) {
		return
	}

	var result model.PingResult
	if err := c.ShouldBindJSON(&result); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	if result.RequestID == "" || result.Executor.NodeID == "" {
		c.JSON(http.StatusBadRequest, gin.H{"error": "request_id and executor.node_id are required"})
		return
	}
	if result.ObservedAt.IsZero() {
		result.ObservedAt = time.Now().UTC()
	}
	if result.ReporterNodeID == "" {
		result.ReporterNodeID = result.Executor.NodeID
	}

	if ok := h.app.Controller.StoreResult(result); !ok {
		c.JSON(http.StatusNotFound, gin.H{"error": "request not found"})
		return
	}

	c.JSON(http.StatusAccepted, gin.H{"status": "accepted"})
}

// getControllerPingRequestResults godoc
//
//	@Summary		Get controller ping request results
//	@Description	Returns a controller request and the results collected for it so far.
//	@Tags			controller
//	@Produce		json
//	@Param			id	path		string	true	"Ping request ID"
//	@Success		200	{object}	ControllerPingRequestResultsResponse
//	@Failure		400	{object}	ErrorResponse
//	@Failure		404	{object}	ErrorResponse
//	@Router			/api/controller/ping-requests/{id}/results [get]
func (h routeHandlers) getControllerPingRequestResults(c *gin.Context) {
	if !ensureController(c, h.app.Controller) {
		return
	}

	requestID := c.Param("id")
	request, ok := h.app.Controller.RequestByID(requestID)
	if !ok {
		c.JSON(http.StatusNotFound, gin.H{"error": "request not found"})
		return
	}

	results := h.app.Controller.ResultsForRequest(requestID)
	c.JSON(http.StatusOK, gin.H{
		"request":    request,
		"collecting": time.Now().UTC().Before(request.CollectUntil),
		"results":    results,
	})
}

func ensureController(c *gin.Context, controllerCtx *controller.Context) bool {
	if controllerCtx != nil {
		return true
	}

	c.JSON(http.StatusBadRequest, gin.H{"error": "node is not running as controller"})
	return false
}

func identityFromHeaders(c *gin.Context) (model.NodeIdentity, bool) {
	nodeID := strings.TrimSpace(c.GetHeader("X-Node-ID"))
	if nodeID == "" {
		return model.NodeIdentity{}, false
	}

	lat, _ := strconv.ParseFloat(strings.TrimSpace(c.GetHeader("X-Node-Lat")), 64)
	lon, _ := strconv.ParseFloat(strings.TrimSpace(c.GetHeader("X-Node-Lon")), 64)

	return model.NodeIdentity{
		NodeID:       nodeID,
		Organization: strings.TrimSpace(c.GetHeader("X-Node-Organization")),
		Role:         model.NodeRole(strings.TrimSpace(c.GetHeader("X-Node-Role"))),
		Location: model.Location{
			Lat: lat,
			Lon: lon,
		},
	}, true
}